refactoring tasks
